clippy fixes

+++ src/service/search/sql.rs
@@ -398,7 +398,7 @@ impl Sql {
             meta.limit = if req_query.size >= 0 {
                 req_query.size as i64
             } else {
-                cfg.limit.query_default_limit * std::cmp::max(1, meta.group_by.len())
+                cfg.limit.query_default_limit * std::cmp::max(1, meta.group_by.len() as i64)
             };
             origin_sql = if meta.order_by.is_empty()
                 && (!sql_mode.eq(&SqlMode::Full)
